move utilities out of the component
