+++ clients/fides-js/src/lib/tcf/utils.ts
@@ -9,9 +9,36 @@ import {
   RecordConsentServedRequest,
 } from "../consent-types";
 import { debugLog } from "../consent-utils";
-import { transformConsentToFidesUserPreference } from "../shared-consent-utils";
+import { transformTcfPreferencesToCookieKeys } from "../cookie";
+import {
+  transformConsentToFidesUserPreference,
+  transformUserPreferenceToBoolean,
+} from "../shared-consent-utils";
+import { generateFidesString } from "../tcf";
 import { FIDES_SYSTEM_COOKIE_KEY_MAP, TCF_KEY_MAP } from "./constants";
 import { decodeFidesString, idsFromAcString } from "./fidesString";
+import {
+  EnabledIds,
+  TCFFeatureRecord,
+  TCFFeatureSave,
+  TcfModels,
+  TCFPurposeConsentRecord,
+  TCFPurposeLegitimateInterestsRecord,
+  TCFPurposeSave,
+  TcfSavePreferences,
+  TCFSpecialFeatureSave,
+  TCFSpecialPurposeSave,
+  TCFVendorConsentRecord,
+  TCFVendorLegitimateInterestsRecord,
+  TCFVendorSave,
+} from "./types";
+
+type TcfSave =
+  | TCFPurposeSave
+  | TCFSpecialPurposeSave
+  | TCFFeatureSave
+  | TCFSpecialFeatureSave
+  | TCFVendorSave;
 
 /**
  * Populates TCF entities with items from both cookie.tcf_consent and cookie.fides_string.
@@ -166,5 +193,141 @@ export const constructTCFNoticesServedProps = (
     tcf_system_consents: minExperience.tcf_system_consent_ids ?? [],
     tcf_system_legitimate_interests:
       minExperience.tcf_system_legitimate_interest_ids ?? [],
+  };
+};
+
+const resolveConsentValueFromTcfModel = (
+  model:
+    | TCFPurposeConsentRecord
+    | TCFPurposeLegitimateInterestsRecord
+    | TCFFeatureRecord
+    | TCFVendorConsentRecord
+    | TCFVendorLegitimateInterestsRecord,
+) => {
+  if (model.current_preference) {
+    return transformUserPreferenceToBoolean(model.current_preference);
+  }
+
+  return transformUserPreferenceToBoolean(model.default_preference);
+};
+
+export const getEnabledIds = (modelList: TcfModels) => {
+  if (!modelList) {
+    return [];
+  }
+  return modelList
+    .map((model) => {
+      const value = resolveConsentValueFromTcfModel(model);
+      return { ...model, consentValue: value };
+    })
+    .filter((model) => model.consentValue)
+    .map((model) => `${model.id}`);
+};
+const transformTcfModelToTcfSave = ({
+  modelList,
+  enabledIds,
+}: {
+  modelList: TcfModels;
+  enabledIds: string[];
+}): TcfSave[] | null => {
+  if (!modelList) {
+    return [];
+  }
+  return modelList.map((model) => {
+    const preference = transformConsentToFidesUserPreference(
+      enabledIds.includes(`${model.id}`),
+    );
+    return {
+      id: model.id,
+      preference,
+    };
+  }) as TcfSave[];
+};
+export const createTcfSavePayload = ({
+  experience,
+  enabledIds,
+}: {
+  experience: PrivacyExperience;
+  enabledIds: EnabledIds;
+}): TcfSavePreferences => {
+  const {
+    tcf_system_consents: consentSystems,
+    tcf_system_legitimate_interests: legintSystems,
+  } = experience;
+  // Because systems were combined with vendors to make the UI easier to work with,
+  // we need to separate them out now (the backend treats them as separate entities).
+  const enabledConsentSystemIds: string[] = [];
+  const enabledConsentVendorIds: string[] = [];
+  const enabledLegintSystemIds: string[] = [];
+  const enabledLegintVendorIds: string[] = [];
+  enabledIds.vendorsConsent.forEach((id) => {
+    if (consentSystems?.map((s) => s.id).includes(id)) {
+      enabledConsentSystemIds.push(id);
+    } else {
+      enabledConsentVendorIds.push(id);
+    }
+  });
+  enabledIds.vendorsLegint.forEach((id) => {
+    if (legintSystems?.map((s) => s.id).includes(id)) {
+      enabledLegintSystemIds.push(id);
+    } else {
+      enabledLegintVendorIds.push(id);
+    }
+  });
+
+  return {
+    purpose_consent_preferences: transformTcfModelToTcfSave({
+      modelList: experience.tcf_purpose_consents,
+      enabledIds: enabledIds.purposesConsent,
+    }) as TCFPurposeSave[],
+    purpose_legitimate_interests_preferences: transformTcfModelToTcfSave({
+      modelList: experience.tcf_purpose_legitimate_interests,
+      enabledIds: enabledIds.purposesLegint,
+    }) as TCFPurposeSave[],
+    special_feature_preferences: transformTcfModelToTcfSave({
+      modelList: experience.tcf_special_features,
+      enabledIds: enabledIds.specialFeatures,
+    }) as TCFSpecialFeatureSave[],
+    vendor_consent_preferences: transformTcfModelToTcfSave({
+      modelList: experience.tcf_vendor_consents,
+      enabledIds: enabledConsentVendorIds,
+    }) as TCFVendorSave[],
+    vendor_legitimate_interests_preferences: transformTcfModelToTcfSave({
+      modelList: experience.tcf_vendor_legitimate_interests,
+      enabledIds: enabledLegintVendorIds,
+    }) as TCFVendorSave[],
+    system_consent_preferences: transformTcfModelToTcfSave({
+      modelList: experience.tcf_system_consents,
+      enabledIds: enabledConsentSystemIds,
+    }) as TCFVendorSave[],
+    system_legitimate_interests_preferences: transformTcfModelToTcfSave({
+      modelList: experience.tcf_system_legitimate_interests,
+      enabledIds: enabledLegintSystemIds,
+    }) as TCFVendorSave[],
+  };
+};
+export const updateCookie = async (
+  oldCookie: FidesCookie,
+  /**
+   * `tcf` and `enabledIds` should represent the same data, where `tcf` is what is
+   * sent to the backend, and `enabledIds` is what the FE uses. They have diverged
+   * because the backend has not implemented separate vendor legint/consents yet.
+   * Therefore, we need both entities right now, but eventually we should be able to
+   * only use one. In other words, `enabledIds` has a field for `vendorsConsent` and
+   * `vendorsLegint` but `tcf` only has `vendors`.
+   */
+  tcf: TcfSavePreferences,
+  enabledIds: EnabledIds,
+  experience: PrivacyExperience,
+): Promise<FidesCookie> => {
+  const tcString = await generateFidesString({
+    tcStringPreferences: enabledIds,
+    experience,
+  });
+  return {
+    ...oldCookie,
+    fides_string: tcString,
+    tcf_consent: transformTcfPreferencesToCookieKeys(tcf),
+    tcf_version_hash: experience.meta?.version_hash,
   };
 };
